Fix incorrect argument for update_with_password
By adding the `resource.extend` line in version 0.9.0, this caused a bug in the `update` method that resulted in `ActiveModel::ForbiddenAttributesError` when trying to update the password.

The solution is to pass in `resource_params` as an argument instead of `params[resource_name]`.

Needs tests here, but I tested it in my app and this fixed the issue.