Make SecurityQuestion ActiveRecord class optional
The SecurityQuestion class is empty, and it's only needed if the `:security_questionable` option is turned on. Therefore, I think it should be up to the user to create the model if they need it.

The problem with automatically creating and requiring the model is that it will override a class by the same name. I have my own implementation of security questions in my app, and was going crazy trying to figure out why the code I put in my `SecurityQuestion` class wasn't being recognized.