Fix redirect issue in password_expired_controller
The code was doing the opposite of what it was intended to do.
It was redirecting to root when the resource is present and when the
resource needs a password change.

Also, this caused `redirect` to be called multiple times in the same
action because calling `redirect_to :root` will not terminate
execution of the action. It will continue to execute the rest of the
code even if the original condition returns true. To stop execution,
you have to add `and return` or use a guard clause, as I did here.

I also added tests for the update action, and configured the dummy
Rails app to support the tests.